requirements.txt
asgiref==3.3.2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.5.0
django-environ==0.9.0
djangorestframework==3.11.1
djangorestframework-simplejwt==5.2.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
